Persist requests sessions for each wiki
When requesting metadata about a specific wiki, use the same session
for all requests. This removes the need to re-establish a connection
every time a new request is sent. This dramatically improves the speed
of processes that require a large number of requests, such as retrieving
a DokuWiki index page or MediaWiki/Fextralife Recent Changes.