Support kwargs passthrough to GET requests
All functions that include GET requests (or call functions that do) now
support kwargs passthrough. The kwargs are passed directly to the GET
request. The "headers" are now passed through this way, instead of all
of these functions having "headers" as an optional parameter.

This supports also passing "timeout" through to requests, to prevent
GET requests hanging indefinitely.
The scripts for profiling wikis and adding redirects now use a default
1 second timeout.