Support adding new languages
When using addredirect.py to add a new language, the script now also
updates other parts of IWB that require updating when adding a new
language.

Unfortunately, this process may disrupt some of the whitespace
surrounding the language dropdown of the settings page.
However, it is isolated to only the language dropdown, as the script
does not touch other parts of the HTML file.