Do not parse lastmod from sitemap
Remove parsing of dates from lastmod when parsing a sitemap.
lastmod is still included as a column in the constructed DataFrame,
but left as an unparsed string.

The lastmod column was never used after retrieving the sitemap,
so there was no need to parse the dates by default.

If the lastmod column is used in the future, the parsing can be done
outside of the retrieve_sitemap function, so that it is not needlessly
performed whenever a sitemap is retrieved.