Keep language in base_url for wiki.gg wikis
For non-English Fandom and wiki.gg wikis, the language should be stored
as part of the base_url instead of as part of the path properties
(content_path and search_path).

There was already code to do this for the content_path of Fandom wikis,
but it did not apply to the search_path.